requirements.txt
# Core dependencies
openai>=1.0.0
whisper-openai>=20231117
PyAudio>=0.2.11
sounddevice>=0.4.6
pyttsx3>=2.90
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0

# Web interface (optional)
streamlit>=1.28.0
plotly>=5.17.0

# Environment management
python-dotenv>=1.0.0

# Audio processing
wave>=0.0.2
pyaudio>=0.2.11

# Alternative TTS services (optional)
azure-cognitiveservices-speech>=1.32.0
elevenlabs>=0.2.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
colorama>=0.4.6
